Replace all variable occurrences in cheat
...not just the first one.

Example: a cheat that runs “tar -czf $folder.tar.gz $folder” would, before this
change, only get the first occurrence of $folder filled in, leaving the rest to
the shell (where $folder would probably not be defined, thus yielding an empty
argument). Now all occurrences of $folder are filled in.