Add two printf cheats to print a screen-wide line
For example, printf 2 with the argument '-' prints the separator line
that’s also used in cheats itself.

I occasionally use the second variant in shell scripts (e. g., in
cheats.sh itself), and having them as cheats is more convenient than
looking them up in the script every time (which is of course the whole
point of cheats). YMMV.