Set local variables instead of replacing $name
By putting the variables and values into the environment proper, we
allow the user to use advanced variable expansions like
${varname:-default value} (#10). The variables are declared local, so
these changes aren’t visible after the cheat has finished running.

Of course, this means that it’s possible for cheat variables to
overwrite internal variables of the function, which is why they’ve been
renamed to make accidental collisions less likely. (For example, the
cheat `find 1` uses a command variable for the -exec command.)

This also means that variable expansions within single quotes are no
longer evaluated, which is why said cheat must be updated.

The help cheat also needs to be updated, because it refers to its own
filename.