Explicitly store cheats invocation to history
I don’t understand the behavior of history -s, but the upshot of it
seems to be, sometimes it replaces the previous entry and sometimes it
appends instead. With this, both the cheats invocation and the cheat
itself will be available in the history afterwards. Don’t ask me why.